tests: Move rate_limiter to integration_rate_limiter.rs
Extract mod rate_limiter from integration.rs into a
standalone test binary integration_rate_limiter.rs.
This follows the pattern established by
integration_cvm.rs.
